refactor: 更新 ReactionRole 模型的 style 屬性類型，將其改為 SmallInteger

+++ starlib/models/mysql.py
@@ -361,7 +361,7 @@ class ReactionRole(AlchemyBasicSchema):
     title: str = Column(String(255))
     description: str | None = Column(String(255), nullable=True)
     emoji: str | None = Column(String(255), nullable=True)
-    style: int | None = Column(Integer, nullable=True)
+    style: int | None = Column(SmallInteger, nullable=True)
 
 
 class Giveaway(BasicSchema, table=True):
